+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "validator-maker",
-  "version": "0.1.1",
+  "version": "0.1.2",
   "description": "Create complex validators from simple ones",
   "keywords": [
     "validator"
@@ -41,6 +41,7 @@
     "prettier": "^3.0.3",
     "typescript": "^5.2.2",
     "vite": "^4.5.0",
+    "vite-plugin-dts": "^3.6.3",
     "vitest": "^0.34.6"
   },
   "license": "MIT"
