+++ vite.config.ts
@@ -1,6 +1,7 @@
 /// <reference types="vitest" />
 
 import { defineConfig } from 'vite'
+import dts from 'vite-plugin-dts'
 import path from 'path'
 
 export default defineConfig({
@@ -11,6 +12,7 @@ export default defineConfig({
       fileName: 'validator-maker',
     },
   },
+  plugins: [dts()],
   test: {
     globals: true,
     mockReset: true,
